% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factbook.R
\name{factbook_cfg}
\alias{factbook_cfg}
\title{Config used internally}
\usage{
factbook_cfg()
}
\value{
list with two slots, one with the base_url of the sqlite
db release of the World Factbook and
one with the dest file where the download ends up locally
}
\description{
This config holds the remote World Factbook database download url
and the corresponding local destination on disk
}
