% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factbook.R
\name{factbook_download}
\alias{factbook_download}
\title{Download the World Factbook as a sqlite3 database}
\usage{
factbook_download(overwrite = FALSE)
}
\arguments{
\item{overwrite}{logical indicating if local db should be overwritten}
}
\value{
logical indicating if the factbook exists locally, invisibly
}
\description{
Download the World Factbook as a sqlite3 database
}
