% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factbook.R
\name{factbook_search_country}
\alias{factbook_search_country}
\title{Full text search for country names}
\usage{
factbook_search_country(search_term)
}
\arguments{
\item{search_term}{token query, phrase query or NEAR query (see http://www.sqlite.org/fts5.html)}
}
\value{
tibble with matching results
}
\description{
Full text search for country names
}
\examples{
 factbook_search_country("vatican")
 factbook_search_country("new")
}
