% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factbook.R
\name{factbook_tibble}
\alias{factbook_tibble}
\title{World Factbook data as a tibble}
\format{A tibble with 261 rows and 13 variables:
\describe{
  \item{id}{row id}
  \item{code}{country code}
  \item{name}{country name}
  \item{area}{total land area}
  \item{area_land}{land area}
  \item{area_water}{water area}
  \item{population}{population}
  \item{population_growth}{population growth}
  \item{birth_rate}{birth rate}
  \item{death_rate}{death rate}
  \item{migration_rate}{migration rate}
  \item{created_at}{created at timestamp}
  \item{updated_at}{updated at timestamp}
}}
\source{
\url{https://github.com/factbook/factbook.sql/releases/download/v0.1/factbook.db}
}
\usage{
factbook_tibble()
}
\description{
A dataset with population, birth_rate and
other attributes for countries around the world.
}
